unify targets for LLVM
