+++ lib/compiler-llvm/src/translator/code.rs
@@ -31,7 +31,7 @@ use target_lexicon::{Architecture, BinaryFormat, OperatingSystem, Triple};
 use wasmer_compiler::WASM_LARGE_FUNCTION_THRESHOLD;
 
 use crate::{
-    abi::{Abi, get_abi},
+    abi::{LLVMAbi, get_abi},
     config::LLVM,
     error::{err, err_nt},
     object_file::{CompiledFunction, load_object_file},
@@ -73,7 +73,7 @@ pub struct FuncTranslator {
     ctx: Context,
     target_triple: Triple,
     target_machines: HashMap<OptimizationStyle, TargetMachine>,
-    abi: Box<dyn Abi>,
+    abi: LLVMAbi,
     binary_fmt: BinaryFormat,
     func_section: String,
     pointer_width: u8,
@@ -331,7 +331,7 @@ impl FuncTranslator {
                 wasm_module,
                 &func,
                 &cache_builder,
-                &*self.abi,
+                &self.abi,
                 self.pointer_width,
                 m0_param,
             ),
@@ -343,7 +343,7 @@ impl FuncTranslator {
             signature_hashes,
             wasm_module,
             symbol_registry,
-            abi: &*self.abi,
+            abi: &self.abi,
             config,
             target_triple: self.target_triple.clone(),
             tags_cache: HashMap::new(),
@@ -2290,7 +2290,7 @@ pub struct LLVMFunctionCodeGenerator<'ctx, 'a> {
     signature_hashes: &'a PrimaryMap<SignatureIndex, SignatureHash>,
     wasm_module: &'a ModuleInfo,
     symbol_registry: &'a dyn SymbolRegistry,
-    abi: &'a dyn Abi,
+    abi: &'a LLVMAbi,
     config: &'a LLVM,
     target_triple: Triple,
     tags_cache: HashMap<i32, BasicValueEnum<'ctx>>,
